Refactor: PaymentsTable and YourRightsBoxChart
