+++ libs/application/templates/parental-leave/src/fields/Review/index.tsx
@@ -32,6 +32,8 @@ import { Payment, Period } from '../../types'
 import { default as format } from 'date-fns/format'
 import Table from '../components/Table'
 import { m } from '../../lib/messages'
+import PaymentsTable from '../PaymentSchedule/PaymentsTable'
+import YourRightsBoxChart from '../Rights/YourRightsBoxChart'
 
 type ValidOtherParentAnswer = 'no' | 'manual' | undefined
 type ValidRadioAnswer = 'yes' | 'no' | undefined
@@ -94,134 +96,6 @@ const Review: FC<FieldBaseProps> = ({ field, application }) => {
       ratio: 50,
     },
   ]
-
-  // TODO: This will come from the formValue
-  const payments: Payment[] = [
-    {
-      date: '2020-12-01T00:00:00.000Z',
-      pensionContribution: 12400,
-      tax: 67000,
-      amount: 389000,
-    },
-    {
-      date: '2021-01-01T00:00:00.000Z',
-      pensionContribution: 14800,
-      tax: 77500,
-      amount: 405300,
-    },
-    {
-      date: '2021-02-01T00:00:00.000Z',
-      pensionContribution: 14800,
-      tax: 77500,
-      amount: 405300,
-    },
-    {
-      date: '2021-03-01T00:00:00.000Z',
-      pensionContribution: 14800,
-      tax: 77500,
-      amount: 405300,
-    },
-    {
-      date: '2021-04-01T00:00:00.000Z',
-      pensionContribution: 11230,
-      tax: 35000,
-      amount: 119000,
-    },
-    {
-      date: '2021-05-01T00:00:00.000Z',
-      pensionContribution: 11230,
-      tax: 35000,
-      amount: 119000,
-    },
-    {
-      date: '2021-06-01T00:00:00.000Z',
-      pensionContribution: 11230,
-      tax: 35000,
-      amount: 119000,
-    },
-    {
-      date: '2021-07-01T00:00:00.000Z',
-      pensionContribution: 11230,
-      tax: 35000,
-      amount: 119000,
-    },
-    {
-      date: '2021-08-01T00:00:00.000Z',
-      pensionContribution: 11230,
-      tax: 35000,
-      amount: 119000,
-    },
-  ]
-
-  // TODO: Copy pasted code from PaymentSchedule/index... will refactor
-  const formatedPayments = payments.map((payment) => {
-    const paymentDate = new Date(payment.date)
-    return {
-      year: format(paymentDate, 'yyyy'),
-      month: format(paymentDate, 'MMMM'),
-      tax: formatIsk(payment.tax),
-      pensionContribution: formatIsk(payment.pensionContribution),
-      amount: formatIsk(payment.amount),
-    }
-  })
-
-  const data = React.useMemo(() => [...formatedPayments], [formatedPayments])
-  const columns = React.useMemo(
-    () => [
-      {
-        Header: formatText(m.salaryLabelYear, application, formatMessage),
-        accessor: 'year', // accessor is the "key" in the data
-      } as const,
-      {
-        Header: formatText(m.salaryLabelMonth, application, formatMessage),
-        accessor: 'month',
-      } as const,
-      {
-        Header: formatText(
-          m.salaryLabelPensionFund,
-          application,
-          formatMessage,
-        ),
-        accessor: 'pensionContribution',
-      } as const,
-      {
-        Header: formatText(m.salaryLabelTax, application, formatMessage),
-        accessor: 'tax',
-      } as const,
-      {
-        Header: formatText(m.salaryLabelPaidAmount, application, formatMessage),
-        accessor: 'amount',
-      } as const,
-    ],
-    [application, formatMessage],
-  )
-
-  const requestRightsAnswer = getValueViaPath(
-    application.answers,
-    'requestRights',
-    undefined,
-  )
-  const giveRightsAnswer = getValueViaPath(
-    application.answers,
-    'giveRights',
-    undefined,
-  )
-
-  const boxChartKeys =
-    requestRightsAnswer === 'yes'
-      ? [
-          { label: '6 personal months', bulletStyle: 'blue' },
-          {
-            label: '1 shared month requested from other parent',
-            bulletStyle: 'greenWithLines',
-          },
-        ]
-      : giveRightsAnswer === 'yes'
-      ? [{ label: '5 personal months', bulletStyle: 'blue' }]
-      : [{ label: '6 personal months', bulletStyle: 'blue' }]
-
-  const numberOfBoxes =
-    requestRightsAnswer === 'yes' ? 7 : giveRightsAnswer === 'yes' ? 5 : 6
 
   return (
     <div>
@@ -457,17 +331,7 @@ const Review: FC<FieldBaseProps> = ({ field, application }) => {
             <Box paddingY={4}>
               <GridRow>
                 <GridColumn span="12/12">
-                  <BoxChart
-                    titleLabel={`Total: ${numberOfBoxes} months *`}
-                    boxes={numberOfBoxes}
-                    calculateBoxStyle={(index) => {
-                      if (index === 6) {
-                        return 'greenWithLines'
-                      }
-                      return 'blue'
-                    }}
-                    keys={boxChartKeys as BoxChartKey[]}
-                  />
+                  <YourRightsBoxChart application={application} />
                 </GridColumn>
               </GridRow>
             </Box>
@@ -498,21 +362,7 @@ const Review: FC<FieldBaseProps> = ({ field, application }) => {
             <Box paddingY={4}>
               <GridRow>
                 <GridColumn span="12/12">
-                  <Table
-                    columns={columns}
-                    data={data}
-                    truncate
-                    showMoreLabel={formatText(
-                      m.salaryLabelShowMore,
-                      application,
-                      formatMessage,
-                    )}
-                    showLessLabel={formatText(
-                      m.salaryLabelShowLess,
-                      application,
-                      formatMessage,
-                    )}
-                  />
+                  <PaymentsTable application={application} />
                 </GridColumn>
               </GridRow>
             </Box>
